// src/common/enums/alert-type.enum.ts
export enum AlertType {
  OVERSPEED = 'OVERSPEED',
  GEOFENCE_ENTER = 'GEOFENCE_ENTER',
  GEOFENCE_EXIT = 'GEOFENCE_EXIT',
  SOS = 'SOS',
  LOW_BATTERY = 'LOW_BATTERY',
  POWER_CUT = 'POWER_CUT',
  VIBRATION = 'VIBRATION',
  DOOR_OPEN = 'DOOR_OPEN',
  ENGINE_ON = 'ENGINE_ON',
  ENGINE_OFF = 'ENGINE_OFF',
  DEVICE_DISCONNECTED = 'DEVICE_DISCONNECTED',
}